Refactoring code, implementing global styles and Tools layout to show on About.
