+++ src/styles/_base.scss
@@ -1,1 +1,21 @@
-// Settings go here
+/* global styles */
+
+@import 'fonts/roboto';
+@import 'typography';
+@import 'links';
+
+:global(body) {
+  position: relative;
+
+  h1 {
+    margin-top: 80px;
+  }
+
+  h2 {
+    margin-bottom: 20px;
+  }
+
+  img {
+    width: 100%;
+  }
+}
