+++ src/styles/links.scss
@@ -1,0 +1,7 @@
+a {
+  text-decoration: none;
+
+  &:hover {
+    text-decoration: none;
+  }
+}
