+++ src/styles/typography.scss
@@ -1,0 +1,19 @@
+/* Typography */
+
+body {
+  font-family: 'Roboto', sans-serif;
+  font-weight: 300;
+}
+
+h1,
+h2,
+h3,
+h4,
+h5,
+h6 {
+  font-weight: 300;
+}
+
+p {
+  font-size: 16px;
+}
